bugFix: router.js pathes updated
