+++ index.js
@@ -17,40 +17,11 @@ app.get('/', function(req, res) {
   // res.write(reqBody);
 });
 
+var resultsCtrl = require("./controllers/results");
+app.use("/results", resultsCtrl);
 
-app.get("/results", function (req, res) {
-  request({
-    url: "http://www.omdbapi.com",
-    qs: {
-      s: req.query.title
-    }
-  }, function (error, response, body) {
-    if (!error && response.statusCode === 200) {
-      var dataObj = JSON.parse(body);
-      res.render("results.ejs", {results: dataObj.Search});
-    }
-  });
-});
-
-app.get("/results", function (req, res) {
-  res.render("results.ejs");
-});
-
-app.get("/details/:id", function (req, res) {
-  request({
-    url: "http://www.omdbapi.com",
-    qs: {
-      i: req.params.id
-      // plot: "long"
-    }
-  }, function (error, response, body) {
-    if (!error && response.statusCode === 200) {
-      var detailsObj = JSON.parse(body);
-      console.log(detailsObj);
-      res.render("details.ejs", {details: detailsObj});
-    }
-  });
-});
+var detailsCtrl = require("./controllers/details");
+app.use("/details", detailsCtrl);
 
 ///////////////////////////////////////////////
 var server = app.listen(process.env.PORT || 3000);
